% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateProteins.R
\name{AnnotateProteins}
\alias{AnnotateProteins}
\title{A function to identify and compile protein enrichment information from a fed peptide subset}
\usage{
AnnotateProteins(
  PeptideVector,
  Treatment,
  FileName,
  Path2FASTA,
  Path2MQev,
  LabelFactor = as.factor(rep(c(rep("Control", 3), rep("Labelled", 3)), 2)),
  EnrichmentFileDir,
  cexSeq = 1,
  ProtPTMs = c("OX", "AC"),
  SeqCharLength = 70,
  GroupPeptides = FALSE,
  verbose = FALSE,
  CorrectLab = TRUE
)
}
\arguments{
\item{PeptideVector}{Character vector with peptide identifiers as delivered by the EnrichmentSet.R function.}

\item{Treatment}{Factor vector with the containing sample information in the same order as samples are outlined in the enrichment file.}

\item{FileName}{Output file name. The output file is a PDF that contains boxplots featuring individual peptide non-corrected LPFs across samples and highlighted peptide sequences in the overall protein sequence provided in the FASTA file.}

\item{Path2FASTA}{FASTA file used for the analysis of the dataset in MaxQuant.}

\item{Path2MQev}{MaxQuant "evidence.txt" file.}

\item{LabelFactor}{Defaults to one treatment and one control, each with a labelled counterpart and triplicated. Needs to be defined as a factor and in the same order as the Treatment vector.}

\item{EnrichmentFileDir}{Parent directory were the enrichment file is contained.}

\item{ProtPTMs}{Character vector with the PTM codes that are part of peptides in the provided dataset.}

\item{SeqCharLength}{Defaults to 70. Defines the length of each line in the output protein sequences.}

\item{GroupPeptides}{Defaults to FALSE. Allows to group all peptides from a single protein into mean protein enrichment.}

\item{verbose}{Defaults to FALSE. When TRUE allows to visualize the progression through the function.}

\item{CorrectLab}{Defaults to TRUE. Corrects labelling percentages using the mean residual "labelling" in non-labelled samples. This ensures that only noise "labelling" percentages with low standard deviation are successfully moved to zero.}
}
\description{
This function allows users to identify the parent proteins from peptide subsets, highlight the coverage of the peptides in the protein sequence, retrieve and visualize the noise-corrected enrichments (non-corrected LPFs) and return an output table that is necessary for the next function in the workflow, which corrects LPFs.
}
\examples{

...
}
\keyword{Dynamics}
\keyword{LCMS}
\keyword{Tracer}
\keyword{kProteomics}
