% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentSet.R
\name{EnrichmentSet}
\alias{EnrichmentSet}
\title{A function to apply thresholds and test statistically peptide enrichment in order to obtain good quality labelled peptide subsets}
\usage{
EnrichmentSet(
  EnrichmentFileDir,
  Treatment = as.factor(c(rep("Control_NL", 3), rep("Control_L", 3),
    rep("Treatment_NL", 3), rep("Treatment_L", 3))),
  LabelFactor = as.factor(rep(c(rep("Control", 3), rep("Labelled", 3)), 2)),
  NLMAX = 0.05,
  LEnrLack = 0.02,
  SigLab = 0.05,
  SigLabControl = 0.05,
  SigLabTreatment = 0.05,
  CorrectLab = TRUE
)
}
\arguments{
\item{EnrichmentFileDir}{Parent directory were the enrichment file is contained.}

\item{Treatment}{Factor vector with the containing sample information in the same order as samples are outlined in the enrichment file.}

\item{LabelFactor}{Defaults to one treatment and one control, each with a labelled counterpart and triplicated. Needs to be defined as a factor and in the same order as the Treatment vector.}

\item{NLMAX}{Defaults to 5% (0.05). Allowed boundaries for noise, "false enrichment" in non-labelled / control samples.}

\item{LEnrLack}{Defaults to 2% (0.02). Below the selected boundary is TRUE lack of enrichment, above is FALSE.}

\item{SigLab}{Defaults to 5% (0.05). Below the selected boundary the FDR corrected P values comparing labeling between control and treatment are significantly different, i.e., TRUE.}

\item{SigLabControl}{Defaults to 5% (0.05). Below the selected boundary the FDR corrected P values comparing non-labelled and labelled control are significantly different, i.e., TRUE.}

\item{SigLabTreatment}{Defaults to 5% (0.05). Below the selected boundary the FDR corrected P values comparing non-labelled and labelled treatment are significantly different, i.e., TRUE.}

\item{CorrectLab}{Defaults to TRUE. Corrects labelling percentages using the mean residual "labelling" in non-labelled samples. This ensures that only noise "labelling" percentages with low standard deviation are successfully moved to zero.}
}
\description{
The statistical filters applied in this section are meant to remove falsely interpreted isotopolog abundances derived from "labelled" controls, this phenomenon can result from peptide coelution and contamination if the heavy isotopolog peaks with different peptides. This in turn results in an increased relative isotope abundance that does not come from the labelling experiment. This special scenario exemplifies the utility of having non-labelled controls in your samples. This function allows users to apply thresholds on residual labelling, noise and multiple parameters that leverage on the quality of the information that is delivered. The function returns a list of peptide subsets that fit the selected criteria and may be fed as input to the subsequent function in order to annotate their parent protein identities in the experimental dataset.
}
\examples{

...
}
\keyword{Dynamics}
\keyword{LCMS}
\keyword{Tracer}
\keyword{kProteomics}
