% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopeEnrichment.R
\name{isotopeEnrichment}
\alias{isotopeEnrichment}
\title{A function to optimize peptide isotopolog intensities and transform them into IsoCorrectoR input files}
\usage{
isotopeEnrichment(
  PyResultsDir,
  returnCSV = TRUE,
  verbose = FALSE,
  rmIsotopologs = 0,
  OptimizeIsotopologNr = FALSE,
  files2correct = NULL,
  AA4correction = NULL,
  AAinterprtFileDir,
  ElementalFileDir,
  ProtPTMs,
  LabelledSamplesNr
)
}
\arguments{
\item{PyResultsDir}{Parent directory were the XX file is contained.}

\item{returnCSV}{Deafults to TRUE. returns the needed IsoCorrectoR MeasurementFile.csv and MoleculeFile.csv to the working directory if set to TRUE.}

\item{verbose}{Defaults to FALSE. Allows to monitor the progression through peptides in order to spot mistakes in the input files if any.}

\item{rmIsotopologs}{Defaults to 0, as an example 3 will keep 3 isotopolog peaks per peptide and discard all others. Allows users to keep a specific number of isotopolog peaks from all input peptides if beyond this number all other peaks should be detrimental to data analysis, e.g., noisy mass spectra.}

\item{OptimizeIsotopologNr}{Defaults to FALSE. Parameter to replace the rmIsotopologs parameter by a customization that calculates the optimal number of isotopolog peaks per peptide taking into account its molecular formula and the enrichment percentage in soluble amino acid precursors. To use it in TRUE mode all the information need to be supplied in the subsequent parameters.}

\item{files2correct}{Defaults to NULL. Needs to be used if OptimizeIsotopologNr is set to TRUE. List of file directories, where each file contains the enrichments in soluble amino acids per treatment evaluated.}

\item{AA4correction}{Defaults to NULL. Needs to be used if OptimizeIsotopologNr is set to TRUE. Character vector with the names of the amino acids to be used for the calculations. The amino acids must be present in the files from the previous parameter.}

\item{AAinterprtFileDir}{File directory to the interpretation file that contains in one column amino acid names and in the second colum the single letter code for those amino acids.}

\item{ElementalFileDir}{Directory to the elemental file to be used in IsoCorrectoR, which contains the natural isotopic abundance (NIA) of chemical elements.}

\item{ProtPTMs}{Character vector with the PTM codes that are part of peptides in the provided dataset.}

\item{LabelledSamplesNr}{Integer reflecting the total number of labelled samples in the dataset.}
}
\description{
This function allows users to grab the output "intensities.tsv" table from isotopeEnrichment.py and transform it into the correct input format for IsoCorrectoR while selecting and optimal number of isotopolog peaks per peptide entry. Optimization of the isotopolog number relies on estimating first how many isotopologs can be expected from the atomic composition of each peptide and their natural isotopic abundance and secondly from the labelling percentage in soluble amino acid pools and the number of labelled amino acid residues in each peptide sequence.
}
\examples{

...
}
\keyword{Dynamics}
\keyword{LCMS}
\keyword{Tracer}
\keyword{kProteomics}
