% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPFcorrection.R
\name{LPFcorrection}
\alias{LPFcorrection}
\title{A function to turn non-corrected into corrected labelled peptide/protein fractions (Corr LPFs)}
\usage{
LPFcorrection(
  InputNonCorrMat,
  files2correct = list.files(path = system.file("extdata", package = "ProtSynthesis"),
    pattern = "Mean_Norm_Factor"),
  AA4correction = c("Serine", "Glycine"),
  AAinterprtFileDir = paste0(system.file("extdata", package = "ProtSynthesis"), "/",
    "AminoAcidNames2SingleLetters.csv"),
  ProtPTMs = c("OX", "AC"),
  CorrectMeans = FALSE,
  EnrBoundary = 1,
  GroupPeptides = T
)
}
\arguments{
\item{files2correct}{Defaults to NULL. List of file directories, where each file contains the enrichments in soluble amino acids per treatment evaluated.}

\item{AA4correction}{Defaults to NULL. Character vector with the names of the amino acids to be used for the calculations. The amino acids must be present in the files from the previous parameter.}

\item{AAinterprtFileDir}{File directory to the interpretation file that contains in one column amino acid names and in the second colum the single letter code for those amino acids.}

\item{ProtPTMs}{Character vector with the PTM codes that are part of peptides in the provided dataset.}

\item{CorrectMeans}{Defaults to FALSE. Allows users to return corrected treatment means instead of individual replicates.}

\item{EnrBoundary}{Defaults to one. Allows to define the boundarie of fractional enrichment accepted to be returned. 1 equals to 100%.}

\item{GroupPeptides}{Defaults to FALSE. Allows to group all peptides from a single protein into mean protein enrichment.}
}
\description{
This function allows users to input their non-corrected LPF matrix from "AnnotateProteins.R" and correct the fractional enrichment in individual peptides using the enrichment percentages in amino acid soluble pools from paired treatments. The function returns the corrected matrix, which can be directly used to calculate fractional synthesis rates.
}
\examples{

...
}
\keyword{Dynamics}
\keyword{LCMS}
\keyword{Tracer}
\keyword{kProteomics}
